+++ thrust/examples/padded_grid_reduction.cu
@@ -80,7 +80,7 @@ int main()
   {
     for (int j = 0; j < n; j++)
     {
-      data[i * N + j] = dist(rng);
+      data[static_cast<std::size_t>(i) * N + j] = dist(rng);
     }
   }
 
@@ -92,7 +92,7 @@ int main()
     std::cout << " ";
     for (int j = 0; j < N; j++)
     {
-      std::cout << data[i * N + j] << " ";
+      std::cout << data[(static_cast<std::size_t>(i) * N) + j] << " ";
     }
     std::cout << "\n";
   }
